+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # @project-version@
-**@project-date-iso@**
 
-- Fixed a bug where an error was fired if you looted a non-weapon or armor item.
+@project-date-iso@
+
+- Ctrl+middle click will now uncheck gear type.
+
+- Tooltip now uses icons and native strings, so no translation is needed for
+that. New colors as well!
+
+- You now whisper a player with ctrl+right click.
+
+- Fixed item linking.